- SQL editor: check if a parser context is available before trying to use it for auto completion. - Fixes a crash when code completion was invoked without a valid connection. - SQL IDE: create a default parser context if no connection could be established (with some restrictions). - db helpers: added int_to_version to convert a simple int version to a GrtVersionRef. - Fixed a few warnings.
- Win: removed the now unnecessary explicit calls to reset refresh handlers for non-db editor wrappers (the others have that already removed or never had it). It's handled now in the base editor.

[TESTS]
- Checked there's no crash anymore with an invalid connection (without a default parser context).
- Checked that the default parser context allows code completion.
- Checked cleanup of non-db object editors on Win.

+++ backend/wbprivate/workbench/wb_context_ui_home.cpp
@@ -605,7 +605,7 @@ void WBContextUI::handle_home_context_menu(const grt::ValueRef &object, const st
     // Internal deletion does not require the prompt
     if (action == "delete_connection_group")
     {
-      std::string text= strfmt(_("Do you really want to delete all the connections in group: %s?"), base::left(group, group.length() -1).c_str());
+      std::string text= strfmt(_("Do you really want to delete all the connections in group: %s?"), base::left(group, (unsigned int)group.length() -1).c_str());
       answer = Utilities::show_warning(_("Delete Connection Group"), text,  _("Delete"), _("Cancel"));
     }
     
@@ -850,7 +850,7 @@ void WBContextUI::handle_home_action(HomeScreenAction action, const grt::ValueRe
       grt::ListRef<db_mgmt_Connection> connections(_wb->get_root()->rdbmsMgmt()->storedConns());
 
       grt::DictRef dict = grt::DictRef::cast_from(object);
-      int to = grt::IntegerRef::cast_from(dict["to"]);
+      int to = (int)grt::IntegerRef::cast_from(dict["to"]);
       if (db_mgmt_ConnectionRef::can_wrap(dict["object"]))
       {
         db_mgmt_ConnectionRef connection = db_mgmt_ConnectionRef::cast_from(dict["object"]);
